updated to move the navbar to a common component
