+++ src/pages/case-study/indrive.tsx
@@ -2,17 +2,13 @@ import CustomerChallenges from "@/components/CustomerCaseStudy/CustomerChallenge
 import CustomerHeader from "@/components/CustomerCaseStudy/CustomerHeader";
 import CustomerTestimonial from "@/components/CustomerCaseStudy/CustomerTestimonial";
 import FooterDev from "@/components/FooterDev/FooterDev";
-import NavbarDev from "@/components/NavbarDev/NavbarDev.component";
-import NavbarStrip from "@/components/NavbarDev/NavbarStrip.component";
-import SummitBanner from "@/components/NavbarDev/SummitBanner.component";
-import { 
+import {
     INDRIVE_CUSTOMER_CHALLENGES, 
     INDRIVE_CUSTOMER_HEADER, 
     INDRIVE_CUSTOMER_HIGHLIGHTS, 
     INDRIVE_CUSTOMER_TESTIMONIAL 
 } from "@/constants/IndriveCustomer.constants";
 import Head from "next/head";
-import { useRouter } from "next/router";
 
 export function getServerSideProps({ resolvedUrl }: { resolvedUrl: string }) {
   return {
@@ -23,12 +19,6 @@ export function getServerSideProps({ resolvedUrl }: { resolvedUrl: string }) {
 }
 
 const IndriveCaseStudyPage = () => {
-  const router = useRouter();
-
-  const handleTryOpenMetadataClick = () => {
-    router.push("/#try-openmetadata");
-  };
-
   return (
     <>
       <Head>
@@ -36,11 +26,6 @@ const IndriveCaseStudyPage = () => {
         <meta name="description" content="See how inDrive tracks and governs 100+ AWS-hosted databases for global teams, enabling faster analytics, automated compliance, and trusted data at scale." />
       </Head>
       <div>
-        <div className="mx-auto fixed top-0 w-full z-[1030]">
-          <SummitBanner />
-          <NavbarDev onClick={handleTryOpenMetadataClick} />
-          <NavbarStrip />
-        </div>
         <div className="mt-20 md:mt-24 lg:mt-32">
           <CustomerHeader
             customerHeader={INDRIVE_CUSTOMER_HEADER}
