+++ src/pages/cve-list.tsx
@@ -1,8 +1,4 @@
 import FooterDev from '@/components/FooterDev/FooterDev';
-import NavbarDev from '@/components/NavbarDev/NavbarDev.component';
-import NavbarStrip from '@/components/NavbarDev/NavbarStrip.component';
-import SummitBanner from "@/components/NavbarDev/SummitBanner.component";
-import { useRouter } from 'next/router';
 import React from 'react';
 
 export function getServerSideProps({ resolvedUrl }: { resolvedUrl: string }) {
@@ -15,20 +11,8 @@ export function getServerSideProps({ resolvedUrl }: { resolvedUrl: string }) {
 }
 
 const CSVList = () => {
-
-  const router = useRouter();
-
-  const handleTryOpenMetadataClick = () => {
-    router.push('/#try-openmetadata');
-  }
-
   return (
     <div id="layoutDefault">
-      <div className="mx-auto fixed top-0 w-full z-[1030]">
-          <SummitBanner />
-          <NavbarDev onClick={handleTryOpenMetadataClick} />
-          <NavbarStrip />
-        </div>
       <div className="right openmetadata-container mx-auto mt-36">
         <h1 className="text-2xl lg:text-4xl font-bold mb-4">
           OpenMetadata Security Vulnerabilities
